Add a base class for transient executioners refs #29149
